Digital max revs in RevMeterWidget can now be boost level dependent.
Fixed Map rotation.

git-svn-id: https://svn.ctdp.net/svn/rfdynhud/trunk@233 1c0d6ce8-a1e3-b74c-a49d-f790e2421cd5

+++ java/src/net/ctdp/rfdynhud/values/IntValue.java
@@ -44,7 +44,7 @@ public class IntValue
         return ( value );
     }
     
-    public final int getFloatValue()
+    public final float getFloatValue()
     {
         return ( value );
     }
